Renamed 'constants' to correctly reflect that the contents are types used for events
